Section services.

+++ feature-main/src/main/java/com/matias/features/ui/main/fragments/home/MainHomeFragment.kt
@@ -78,7 +78,7 @@ class MainHomeFragment :
 			offscreenPageLimit = 2
 			clipToPadding = false
 			activity?.let {
-				this.setPadding(dpToPx(it, 12), 0, dpToPx(it, 12), 0)
+				this.setPadding(dpToPx(it, 12F).toInt(), 0, dpToPx(it, 12F).toInt(), 0)
 //				this.pageMargin = dpToPx(it, 4)
 			}
 		}
